config/nvim/which-key: Adjust window management binding
This is still under the experimentation but here is
the summary of the window management binding:

<space>wv: create a vertical window
<space>wh: create a holizontal window
<space>wf: Taller the window's height
<space>ws: Shorter the window's height
<space>ww: Widen the window's width
<space>wn: Narrow the window's width

I tried wt, instead of wf, for taller binding
but t is a bit harder to type than f.

It's still under the experimentation and I might
come back with the better binding, hopefully, in
the future.